YH-1002: Replace react-slick usage with shared Slider component
