+++ src/shared/ui/Slider/index.ts
@@ -1,0 +1,3 @@
+import { Slider } from './Slider';
+
+export { Slider };
