+++ src/widgets/Landing/HistoryBlock/ui/HistorySliderBlock/HistorySliderBlock.tsx
@@ -1,14 +1,12 @@
 import { useTranslation } from 'react-i18next';
-import Slider from 'react-slick';
 
 import { i18Namespace } from '@/shared/config/i18n';
 import { Landing } from '@/shared/config/i18n/i18nTranslations';
+import { Slider } from '@/shared/ui/Slider';
 
 import { historySlides, sliderSettings } from '../../model/constants';
 import { HistorySlide } from '../HistorySlide/HistorySlide';
 
-import 'slick-carousel/slick/slick-theme.css';
-import 'slick-carousel/slick/slick.css';
 import styles from './HistorySliderBlock.module.css';
 
 export const HistorySliderBlock = () => {
